+++ lib/pages/home_detail_page.dart
@@ -1,0 +1,23 @@
+import 'package:carousel_app/models/catalog.dart';
+import 'package:flutter/material.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+class HomeDetailsPage extends StatelessWidget {
+  final Item catalog;
+  const HomeDetailsPage({Key? key, required this.catalog})
+      : assert(catalog != null),
+        super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: SafeArea(
+        child: Column(
+          children: [
+            Image.network(catalog.image),
+          ],
+        ).centered(),
+      ),
+    );
+  }
+}
