+++ lib/pages/home_page.dart
@@ -1,6 +1,6 @@
 import 'package:carousel_app/models/catalog.dart';
-import 'package:carousel_app/widgets/drawer.dart';
-import 'package:carousel_app/widgets/item_widget.dart';
+import 'package:carousel_app/pages/home_widgets/catalog_header.dart';
+import 'package:carousel_app/pages/home_widgets/catalog_list.dart';
 import 'package:carousel_app/widgets/themes.dart';
 import 'package:flutter/material.dart';
 //import 'package:flutter/material.dart';
@@ -43,119 +43,26 @@ class _HomePageState extends State<HomePage> {
     return Scaffold(
         backgroundColor: MyTheme.creamColor,
         body: SafeArea(
-          child: Container(
-            padding: Vx.m32,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                CatalogHeader(),
-                if (CatalogModel.items != null && CatalogModel.items.isNotEmpty)
-                  CatalogList().expand()
-                else
-                  const Center(
-                    child: CircularProgressIndicator(),
-                  )
-              ],
+          child: Scrollbar(
+            thumbVisibility: true,
+            child: Container(
+              padding: Vx.m32,
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  const CatalogHeader(),
+                  if (CatalogModel.items != null &&
+                      CatalogModel.items.isNotEmpty)
+                    const CatalogList().expand()
+                  else
+                    const CircularProgressIndicator()
+                        .centered()
+                        .py16()
+                        .expand(),
+                ],
+              ),
             ),
           ),
         ));
-  }
-}
-
-class CatalogHeader extends StatelessWidget {
-  const CatalogHeader({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        "Catalog App".text.xl5.bold.make(),
-        "Trending Products".text.xl2.make(),
-        Container(
-          height: 20,
-        ),
-        const Divider(
-          color: Colors.black,
-          height: 10,
-        )
-      ],
-    );
-  }
-}
-
-class CatalogList extends StatelessWidget {
-  const CatalogList({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return ListView.builder(
-      shrinkWrap: true,
-      itemBuilder: (context, index) {
-        final catalog = CatalogModel.items[index];
-        return CatalogItem(catalog: catalog);
-      },
-      itemCount: CatalogModel.items.length,
-    );
-  }
-}
-
-class CatalogItem extends StatelessWidget {
-  final Item catalog;
-  const CatalogItem({Key? key, required this.catalog})
-      : assert(catalog != null),
-        super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return VxBox(
-        child: Row(
-      children: [
-        CatalogImage(image: catalog.image),
-        Expanded(
-            child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            catalog.name.text.lg.bold.make(),
-            catalog.desc.text.textStyle(context.captionStyle).make(),
-            10.heightBox,
-            ButtonBar(
-              buttonPadding: EdgeInsets.zero,
-              alignment: MainAxisAlignment.spaceBetween,
-              children: [
-                "\$${catalog.price}".text.bold.xl.make(),
-                ElevatedButton(
-                  onPressed: () {},
-                  style: ButtonStyle(
-                      backgroundColor: MaterialStateProperty.all(
-                          const Color.fromARGB(255, 40, 39, 42)),
-                      shape: MaterialStateProperty.all(StadiumBorder())),
-                  child: "Buy".text.make(),
-                ).pOnly(right: 20.0),
-              ],
-            )
-          ],
-        ))
-      ],
-    )).white.rounded.square(150).make().py16();
-  }
-}
-
-class CatalogImage extends StatelessWidget {
-  final String image;
-
-  const CatalogImage({super.key, required this.image});
-
-  @override
-  Widget build(BuildContext context) {
-    return Image.network(image)
-        .box
-        .rounded
-        .p16
-        .color(Colors.white)
-        .make()
-        .p16()
-        .w32(context);
   }
 }
