+++ lib/pages/home_widgets/catalog_header.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+class CatalogHeader extends StatelessWidget {
+  const CatalogHeader({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        "Catalog App".text.xl5.bold.make(),
+        "Trending Products".text.xl2.make(),
+        Container(
+          height: 20,
+        ),
+        const Divider(
+          color: Colors.black,
+          height: 10,
+        )
+      ],
+    );
+  }
+}
