+++ lib/pages/home_widgets/catalog_image.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+class CatalogImage extends StatelessWidget {
+  final String image;
+
+  const CatalogImage({super.key, required this.image});
+
+  @override
+  Widget build(BuildContext context) {
+    return Image.network(image)
+        .box
+        .rounded
+        .p4
+        .color(Colors.white)
+        .make()
+        .p16()
+        .w32(context);
+  }
+}
